breakup files into units
